+++ Editor/Test/JSONSerializer.cs
@@ -27,15 +27,17 @@ namespace EPPZ.Persistence.Editor.Test
 
 
 		EPPZ.Persistence.JSONSerializer serializer;
-		string testFolderPath;
-		string tempFolderPath;
+
+		string testFolderPath; // Absolute
+		string tempFolderPath; // Absolute
+		string resourcesFolderPath; // Absolute
 
 		Payload alpha, beta, gamma;
 		Entity first, second, third, fourth;
 		string first_JSON, second_JSON, third_JSON, fourth_JSON;
 
 
-		[SetUp]
+		[OneTimeSetUp]
 		public void Setup()
 		{
 			serializer = new EPPZ.Persistence.JSONSerializer();
@@ -85,13 +87,15 @@ namespace EPPZ.Persistence.Editor.Test
 				"Fourth"
 			);
 
-					
-
 			// JSON representation (escaped double quotes).
 			first_JSON = "{\"ID\":1,\"name\":\"First\",\"payloads\":[{\"ID\":1,\"name\":\"Alpha\",\"data\":[0,1,2,3,4,5,6,7,8,9]},{\"ID\":2,\"name\":\"Beta\",\"data\":[0,1,2,3,4,5,6,7,8,9]},{\"ID\":3,\"name\":\"Gamma\",\"data\":[0,1,2,3,4,5,6,7,8,9]}]}";
 			second_JSON = "{\"ID\":2,\"name\":\"Second\",\"payloads\":[{\"ID\":2,\"name\":\"Beta\",\"data\":[0,1,2,3,4,5,6,7,8,9]},{\"ID\":3,\"name\":\"Gamma\",\"data\":[0,1,2,3,4,5,6,7,8,9]}]}";
 			third_JSON = "{\"ID\":3,\"name\":\"Third\",\"payloads\":[{\"ID\":3,\"name\":\"Gamma\",\"data\":[0,1,2,3,4,5,6,7,8,9]}]}";
 			fourth_JSON = "{\"ID\":4,\"name\":\"Fourth\",\"payloads\":[]}";
+
+			// _RecreateTestFiles();
+			Files.CreateFolderIfNotExist(tempFolderPath);
+			_Setup_Resources();
 		}
 
 		void _RecreateTestFiles()
@@ -102,9 +106,9 @@ namespace EPPZ.Persistence.Editor.Test
 			serializer.SerializeObjectToFile(fourth, testFolderPath + "fourth", Mode.Pretty);	
 		}
 
-		[TearDown]
-		public void TeaerDown()
-		{ }
+		[OneTimeTearDown]
+		public void TearDown()
+		{ _TearDown_Resources(); }
 
 
 	#region Model
@@ -241,40 +245,34 @@ namespace EPPZ.Persistence.Editor.Test
 		[Test]
 		public void ObjectToFile_Pretty()
 		{
- 			if (!Directory.Exists(tempFolderPath))
-            { Directory.CreateDirectory(tempFolderPath); }
-
 			serializer.ObjectToFile(first, tempFolderPath+"first_test_pretty", Mode.Pretty);
-			Assertation.AreFilesEqual(
+			Files.AreEqual(
 				testFolderPath+"first.json",
 				tempFolderPath+"first_test_pretty.json"
 			);
 
 			serializer.ObjectToFile(second, tempFolderPath+"second_test_pretty", Mode.Pretty);
-			Assertation.AreFilesEqual(
+			Files.AreEqual(
 				testFolderPath+"second.json",
 				tempFolderPath+"second_test_pretty.json"
 			);
 
 			serializer.ObjectToFile(third, tempFolderPath+"third_test_pretty", Mode.Pretty);
-			Assertation.AreFilesEqual(
+			Files.AreEqual(
 				testFolderPath+"third.json",
 				tempFolderPath+"third_test_pretty.json"
 			);
 
 			serializer.ObjectToFile(fourth, tempFolderPath+"fourth_test_pretty", Mode.Pretty);
-			Assertation.AreFilesEqual(
+			Files.AreEqual(
 				testFolderPath+"fourth.json",
 				tempFolderPath+"fourth_test_pretty.json"
 			);
 		}
 
 		[Test]
-		public void ObjectToFile()
-		{
- 			if (!Directory.Exists(tempFolderPath))
-            { Directory.CreateDirectory(tempFolderPath); }
-			
+		public void ObjectToFile_Default()
+		{			
 			serializer.ObjectToFile(first, tempFolderPath+"first_test");
 			Assert.AreEqual(
 				first_JSON,
@@ -302,6 +300,120 @@ namespace EPPZ.Persistence.Editor.Test
 
 	#endregion
 	
+
+	#region Resource
+
+		void _Setup_Resources()
+		{
+			resourcesFolderPath = Application.dataPath + "/Resources/";
+			Files.CreateFolder(resourcesFolderPath);
+
+			Files.Copy(
+				testFolderPath + "first.json",
+				resourcesFolderPath + "first.json"
+			);
+
+			Files.Copy(
+				testFolderPath + "second.json",
+				resourcesFolderPath + "second.json"
+			);
+
+			Files.Copy(
+				testFolderPath + "third.json",
+				resourcesFolderPath + "third.json"
+			);
+
+			Files.Copy(
+				testFolderPath + "fourth.json",
+				resourcesFolderPath + "fourth.json"
+			);
+
+			AssetDatabase.ImportAsset("Assets/Resources/first.json");
+			AssetDatabase.ImportAsset("Assets/Resources/second.json");
+			AssetDatabase.ImportAsset("Assets/Resources/third.json");
+			AssetDatabase.ImportAsset("Assets/Resources/fourth.json");
+
+		}
+
+		void _TearDown_Resources()
+		{
+			Files.DeleteAsset("Assets/Resources/first.json");
+			Files.DeleteAsset("Assets/Resources/second.json");
+			Files.DeleteAsset("Assets/Resources/third.json");
+			Files.DeleteAsset("Assets/Resources/fourth.json");
+
+			Files.DeleteAsset("Assets/Resources/first_test_pretty.json");
+			Files.DeleteAsset("Assets/Resources/second_test_pretty.json");
+			Files.DeleteAsset("Assets/Resources/third_test_pretty.json");
+			Files.DeleteAsset("Assets/Resources/fourth_test_pretty.json");
+
+			Files.DeleteAsset("Assets/Resources/first_test.json");
+			Files.DeleteAsset("Assets/Resources/second_test.json");
+			Files.DeleteAsset("Assets/Resources/third_test.json");
+			Files.DeleteAsset("Assets/Resources/fourth_test.json");
+
+			Files.DeleteFolderIfEmpty(resourcesFolderPath);
+		}
+
+		[Test]
+		public void ResourceToObject()
+		{
+			Assert.AreEqual(
+				first,
+				serializer.ResourceToObject<Entity>("first")
+			);
+			
+			Assert.AreEqual(
+				second,
+				serializer.ResourceToObject<Entity>("second")
+			);
+			
+			Assert.AreEqual(
+				third,
+				serializer.ResourceToObject<Entity>("third")
+			);
+			
+			Assert.AreEqual(
+				fourth,
+				serializer.ResourceToObject<Entity>("fourth")
+			);
+		}
+
+		[Test]
+		public void ObjectToResource()
+		{			
+			// Write to `Assets/Resources`, import, then read back as resource (using `UnityEngine.Resources`).
+			serializer.ObjectToFile(first, resourcesFolderPath + "first_test");
+			AssetDatabase.ImportAsset("Assets/Resources/first_test.json");
+			Assert.AreEqual(
+				first_JSON,
+				(Resources.Load("first_test") as TextAsset).text
+			);
+
+			serializer.ObjectToFile(second, resourcesFolderPath + "second_test");
+			AssetDatabase.ImportAsset("Assets/Resources/second_test.json");
+			Assert.AreEqual(
+				second_JSON,
+				(Resources.Load("second_test") as TextAsset).text
+			);
+
+			serializer.ObjectToFile(third, resourcesFolderPath + "third_test");
+			AssetDatabase.ImportAsset("Assets/Resources/third_test.json");
+			Assert.AreEqual(
+				third_JSON,
+				(Resources.Load("third_test") as TextAsset).text
+			);
+
+			serializer.ObjectToFile(fourth, resourcesFolderPath + "fourth_test");
+			AssetDatabase.ImportAsset("Assets/Resources/fourth_test.json");
+			Assert.AreEqual(
+				fourth_JSON,
+				(Resources.Load("fourth_test") as TextAsset).text
+			);
+		}
+
+	#endregion
+
 
 	}
 }